refactor: rename entities
