+++ src/entities/LoanData.ts
